refactor: 빌더 패턴으로 수정 및 패키지 명 변경
