+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/user/auth/CustomSuccessHandler.java
@@ -3,7 +3,7 @@ package com.mars.NangPaGo.domain.user.auth;
 import com.mars.NangPaGo.domain.user.dto.RefreshTokenDto;
 import com.mars.NangPaGo.domain.user.repository.RefreshTokenRepository;
 import com.mars.NangPaGo.domain.user.util.JwtUtil;
-import com.mars.NangPaGo.domain.user.vos.CustomOAuth2User;
+import com.mars.NangPaGo.domain.user.vo.CustomOAuth2User;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
