+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/user/factory/userinfo/GoogleUserInfo.java
@@ -3,9 +3,7 @@ package com.mars.NangPaGo.domain.user.factory.userinfo;
 import java.util.Map;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import lombok.ToString;
 
-@ToString
 @RequiredArgsConstructor
 public class GoogleUserInfo implements OAuth2UserInfo {
