+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/user/service/CustomOAuth2UserService.java
@@ -6,7 +6,7 @@ import com.mars.NangPaGo.domain.user.entity.User;
 import com.mars.NangPaGo.domain.user.factory.OAuth2UserInfoFactory;
 import com.mars.NangPaGo.domain.user.factory.userinfo.OAuth2UserInfo;
 import com.mars.NangPaGo.domain.user.repository.UserRepository;
-import com.mars.NangPaGo.domain.user.vos.CustomOAuth2User;
+import com.mars.NangPaGo.domain.user.vo.CustomOAuth2User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
